getMeta(context)
Meta information is particular to a context, so the cached meta information on a cell should be invalidated if you're trying to use it for a different context. Additionally, this means that the meta-object contains all the information necessary to resolve nocks, kicks, etc, which we will soon exploit in polyglot-land via findMetaObject().