Start a skeleton of the serf.
We can't access any of the nock-language internals through the
polyglot API, so we must expose jam/cue as bindings. This adds
a kinda fake test to show that they work.

This also starts a skeleton of a serf, with the "ability" to
at least construct the `[%play ~]` which should be sent to the
king to signal that the serf is up.