Changing what core we look up on the parentCore fixes +puck.
When trying to register the +puck:ed:crypto jet, we were previously
failing with some really weird battery lookup issues. I belive this
was because when we were trying to look up the core in the
parentCore instead of parentCore in parentCore. Switching this
fixes registration of zuse jets.

The question though is, if this was actually what was broken, how
did anything actually work?